requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.4
django-allauth==0.61.1
idna==3.7
lxml==5.2.1
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
